Doc changes and Number Features added to Clearing Ops
